fix(create-post-page): fix pinia loading error
