--- apps/ourvoice-app/src/components/common/Dropdown.vue
@@ -1,113 +1,0 @@
-<template>
-    <div class="flex">
-        <div class="flex-1">
-            <Listbox v-model="orderBy">
-                <div class="relative mt-1">
-                    <ListboxButton
-                        class="relative w-full cursor-default rounded-lg bg-white py-2 pl-3 pr-10 text-left shadow-md focus:outline-none focus-visible:border-indigo-500 focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 focus-visible:ring-offset-2 focus-visible:ring-offset-orange-300 sm:text-sm">
-                        <span class="block truncate">{{ orderBy.name }}</span>
-                        <span class="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
-                            <ChevronUpDownIcon class="h-5 w-5 text-gray-400" aria-hidden="true" />
-                        </span>
-                    </ListboxButton>
-
-                    <transition leave-active-class="transition duration-100 ease-in" leave-from-class="opacity-100"
-                        leave-to-class="opacity-0">
-                        <ListboxOptions
-                            class="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                            <ListboxOption v-slot="{ active, selected }" v-for="opt in orderOptions" :key="opt.name"
-                                :value="opt" as="template">
-                                <li :class="[
-                                    active ? 'bg-amber-100 text-amber-900' : 'text-gray-900',
-                                    'relative cursor-default select-none py-2 pl-10 pr-4',
-                                ]">
-                                    <span :class="[
-                                        selected ? 'font-medium' : 'font-normal',
-                                        'block truncate',
-                                    ]">{{ opt.name }}</span>
-                                    <span v-if="selected"
-                                        class="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
-                                        <CheckIcon class="h-5 w-5" aria-hidden="true" />
-                                    </span>
-                                </li>
-                            </ListboxOption>
-                        </ListboxOptions>
-                    </transition>
-                </div>
-            </Listbox>
-        </div>
-
-        <div class="flex-1">
-            <Listbox v-model="filterBy">
-                <div class="relative mt-1">
-                    <ListboxButton
-                        class="relative w-full cursor-default rounded-lg bg-white py-2 pl-3 pr-10 text-left shadow-md focus:outline-none focus-visible:border-indigo-500 focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 focus-visible:ring-offset-2 focus-visible:ring-offset-orange-300 sm:text-sm">
-                        <span class="block truncate">{{ filterBy.name }}</span>
-                        <span class="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
-                            <ChevronUpDownIcon class="h-5 w-5 text-gray-400" aria-hidden="true" />
-                        </span>
-                    </ListboxButton>
-
-                    <transition leave-active-class="transition duration-100 ease-in" leave-from-class="opacity-100"
-                        leave-to-class="opacity-0">
-                        <ListboxOptions
-                            class="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                            <ListboxOption v-slot="{ active, selected }" v-for="opt in filterOptions" :key="opt.name"
-                                :value="opt" as="template">
-                                <li :class="[
-                                    active ? 'bg-amber-100 text-amber-900' : 'text-gray-900',
-                                    'relative cursor-default select-none py-2 pl-10 pr-4',
-                                ]">
-                                    <span :class="[
-                                        selected ? 'font-medium' : 'font-normal',
-                                        'block truncate',
-                                    ]">{{ opt.name }}</span>
-                                    <span v-if="selected"
-                                        class="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
-                                        <CheckIcon class="h-5 w-5" aria-hidden="true" />
-                                    </span>
-                                </li>
-                            </ListboxOption>
-                        </ListboxOptions>
-                    </transition>
-                </div>
-            </Listbox>
-        </div>
-    </div>
-
-    <button @click="handleSubmit"
-        class="bg-blue-500 hover:bg-blue-600 text-white px-6 py-2 rounded-lg shadow-md flex items-center transition duration-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-indigo-300 disabled:cursor-not-allowed"
-        data-cy="create-post-submit-button">
-        Apply
-    </button>
-</template>
-
-<script lang="ts" setup>
-import { ref } from 'vue'
-import {
-    Listbox,
-    ListboxButton,
-    ListboxOptions,
-    ListboxOption,
-} from '@headlessui/vue'
-import { CheckIcon, ChevronUpDownIcon } from '@heroicons/vue/20/solid'
-
-const orderOptions = [
-    { name: 'Latest First' },
-    { name: 'Most Commented' },
-    { name: 'Most Voteup' },
-    { name: 'Most Votedown' },
-]
-
-const filterOptions = [
-    { name: 'Latest' },
-    { name: 'All Time' },
-]
-const orderBy = ref(orderOptions[0])
-const filterBy = ref(filterOptions[0])
-
-const handleSubmit = () => {
-    console.log(orderBy.value)
-    console.log(filterBy.value)
-}
-</script>
