+++ apps/ourvoice-app/src/components/common/TabsWrapper.vue
@@ -3,10 +3,10 @@
     <div class="tabs max-w-lg text-center m-auto">
       <ul class="tabs_header flex list-none">
         <li
-          class="bg-slate-400 mr-2 p-1 rounded-t-lg text-white cursor-pointer"
-          :class="{ 'bg-indigo-500': activeTab === tabTitle }"
           v-for="tabTitle in tabTitles"
           :key="tabTitle"
+          class="mr-2 p-1 rounded-t-lg text-white cursor-pointer"
+          :class="[activeTab === tabTitle ? 'bg-indigo-500' : 'bg-slate-400']"
           @click="activeTab = tabTitle"
         >
           {{ tabTitle }}
@@ -26,7 +26,7 @@ export default {
         return slot.props?.title
       })
     )
-    const activeTab = ref(tabTitles.value[0])
+    const activeTab = ref(tabTitles?.value[0])
 
     provide('activeTab', activeTab)
     return {
